email service updated

+++ src/main/kotlin/org/generis/business/invoice/dto/CreateInvoiceDto.kt
@@ -11,5 +11,5 @@ data class CreateInvoiceDto(
     var useCompanyCurrency: Boolean = false,
     var dueDate: String,
     var tax: Double? = 0.00,
-    var discount: Double? =0.00
+    var discount: Int? =0
 )
